cleanup and optimisations
switching to new settings screen
