+++ e2e/functional/functionalTest.spec.js
@@ -28,7 +28,7 @@ describe('functional test', () => {
     describe('cgm simulator screen', () => {
         functionalityTests.cgmSimulatorScreen(test);
     });
-    describe('delivery limits screen', () => {
+    describe.skip('delivery limits screen', () => {
         functionalityTests.deliveryLimitsScreen(test);
     });
 });
