+++ src/config/enUSText.js
@@ -113,10 +113,13 @@ var SettingsScreen = {
     DeliveryLimits: 'Delivery Limits',
     InsulinModel: 'Insulin Model',
     ClosedLoop: 'Closed Loop',
-    AddCGM: 'Add CGM',
+    //AddCGM: 'Add CGM',
+    AddCGM: 'Add CGM\nTap here to set up a CGM',
     DeleteCGMData: 'Delete CGM Data',
     DeleteCGM: 'Delete CGM',
-    AddPump: 'Add Pump',
+    //AddPump: 'Add Pump',
+    AddPump: 'Add Pump\nTap here to set up a pump',
+    SimulatorPump: 'Simulator Small',
     DeletePump: 'Delete Pump',
     DeletePumpData: 'Delete Pump Data',
     PreMeal: 'Pre-Meal',
