+++ tests/functionality/settingsScreen.tests.js
@@ -1,25 +1,25 @@
 var settingsScreen = (test) => {
     var settingsScreen;
-    it('can open the settings', async () => {
+    it('can open', async () => {
         settingsScreen = await test.OpenSettingsScreen();
     });
     it('set to closed loop', async () => {
-        await settingsScreen.SetClosedLoop();
+        await settingsScreen.ClosedLoop();
     });
     it('set to open loop', async () => {
-        await settingsScreen.SetOpenLoop();
+        await settingsScreen.OpenLoop();
     });
-    describe('issue report', () => {
-        var screen;
-        it('can be opened', async () => {
-            screen = await settingsScreen.OpenIssueReportScreen();
-        });
-        it('can closed', async () => {
-            await screen.Close();
-        });
+    it('support opened', async () => {
+        await settingsScreen.OpenSupport();
     });
-    it('can close the settings', async () => {
-        await settingsScreen.Close();
+    it(' supportclosed', async () => {
+        await settingsScreen.CloseSupport();
+    });
+    it.skip('therapy settings opened', async () => {
+        await settingsScreen.OpenTherapySettings();
+    });
+    it('can close', async () => {
+        await settingsScreen.BackToHome();
     });
 };
